refactored into multiple files and plugins, started on camera control
